% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rptFuncs.r
\name{binBreaker}
\alias{binBreaker}
\title{Bin Breaker}
\usage{
binBreaker(x, nbin)
}
\arguments{
\item{x}{vector to find cuts for}

\item{nbin}{number of bins}
}
\value{
list with elements Breaks and Labels
}
\description{
Generate break points for use with cut()
and range labels based on sample max and min
}
\details{
helpful in plot making plots
Labels use the acutal max and min rather than +/- 1% used by cut()
}

