% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpFuncs.r
\name{epanetDefaultOptions}
\alias{epanetDefaultOptions}
\title{Epanet Default Options}
\usage{
epanetDefaultOptions()
}
\description{
A list of Epanet's default options
}
\details{
Provides a named list in the form of OPTION = default_value where the 
values are taken from pages 152-154 of the manual.
}
\examples{
epanetDefaultOptions() 
}
\references{
Rossman, L. A. (2000). Epanet 2 users manual. US EPA, Cincinnati, Ohio.
http://nepis.epa.gov/Adobe/PDF/P1007WWU.pdf
}
