% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanetReader.r
\docType{package}
\name{epanetReader-package}
\alias{epanetReader}
\alias{epanetReader-package}
\title{epanetReader: Read Epanet Files into R}
\description{
Reads water network simulation data in 'Epanet' text-based '.inp' and '.rpt' formats into R. Also reads results from 'Epanet-msx'. Provides basic summary information and plots. The README file has a quick introduction. See \url{https://www.epa.gov/water-research/epanet} for more information on the Epanet software for modeling hydraulic and water quality behavior of water piping systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bradleyjeck/epanetReader}
}

}
\author{
\strong{Maintainer}: Bradley Eck \email{brad@bradeck.net}

}
\keyword{internal}
