% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expandedLinkTable-s3.r
\name{expandedLinkTable}
\alias{expandedLinkTable}
\title{Expanded Link Table}
\usage{
expandedLinkTable(Links, Coordinates)
}
\arguments{
\item{Links}{data frame of Pipes, Pumps or Valves of from epanet.inp}

\item{Coordinates}{table of epanet.inp}
}
\description{
Add coordinates of nodes to data frame
of pipes, pumps, or valves from an epanet.inp object
}

