% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.inp-s3.r
\name{is.epanet.inp}
\alias{is.epanet.inp}
\title{Check if an object as class 'epanet.inp'}
\usage{
is.epanet.inp(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Check if an object as class 'epanet.inp'
}
