% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.rpt-s3.r
\name{is.epanet.rpt}
\alias{is.epanet.rpt}
\title{Check if an object has class 'epanet.rpt'}
\usage{
is.epanet.rpt(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Check if an object has class 'epanet.rpt'
}
