% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanetmsx.rpt-s3.r
\name{is.epanetmsx.rpt}
\alias{is.epanetmsx.rpt}
\title{Check if an object has class 'epanetmsx.rpt'}
\usage{
is.epanetmsx.rpt(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Check if an object has class 'epanetmsx.rpt'
}
