% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandedLinkTable-s3.r
\name{is.expandedLinkTable}
\alias{is.expandedLinkTable}
\title{Check if an object has class 'expandedLinkTable'}
\usage{
is.expandedLinkTable(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Check if an object has class 'expandedLinkTable'
}
