% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline-s3.r
\name{is.sparkline}
\alias{is.sparkline}
\title{Check if an object has class 'sparkline'}
\usage{
is.sparkline(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Check if an object has class 'sparkline'
}
