% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparklineTable-s3.r
\name{is.sparklineTable}
\alias{is.sparklineTable}
\title{Check if an object has class 'sparklineTable'}
\usage{
is.sparklineTable(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Check if an object has class 'sparklineTable'
}
