% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Net1-data.r
\docType{data}
\name{Net1}
\alias{Net1}
\title{Epanet's Net1 Example}
\format{An object of class \code{epanet.inp} created by \link{read.inp}.}
\source{
http://www.epa.gov/sites/production/files/2014-06/en2setup_0.exe
}
\usage{
Net1
}
\description{
A dataset created by reading the Net1.inp file
distributed with Epanet using this package's
read.inp() function.
}
\keyword{datasets}

