% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Net1rpt-data.r
\docType{data}
\name{Net1rpt}
\alias{Net1rpt}
\title{Epanet's Net1 Example}
\format{An object of class \code{epanet.rpt} created by \link{read.inp}.}
\usage{
Net1rpt
}
\description{
A dataset created by reading the Net1.inp file
distributed with Epanet using this package's
read.inp() function.
}
