% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epanet.rpt-s3.r
\name{plot.epanet.rpt}
\alias{plot.epanet.rpt}
\title{Plot Simulation Results}
\usage{
\method{plot}{epanet.rpt}(x, inp, Timestep = "0:00:00", juncQty = "Demand",
  linkQty = "Velocity", legend1.locn = "topright",
  legend2.locn = "topleft", ...)
}
\arguments{
\item{x}{epanet.rpt object}

\item{inp}{epanet.inp object associated with x}

\item{Timestep}{string indicating the time at which to plot}

\item{juncQty}{string specifying which column of node results table
to plot for junctions}

\item{linkQty}{string specifying which column of link results table
to plot}

\item{legend1.locn}{placement of legend for network elements}

\item{legend2.locn}{placement of legend for junction and link quantities}

\item{...}{further arguments passed to plot}
}
\description{
Plots simulation results for a single time step in map form
}
\details{
juncQty is scaled over only junction result types.  linkQty is
        scaled over all of the link types
}

