% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanetmsx.rpt-s3.r
\name{plot.epanetmsx.rpt}
\alias{plot.epanetmsx.rpt}
\title{Plot method for epanetmsx.rpt}
\usage{
\method{plot}{epanetmsx.rpt}(x, elementType = "Nodes", ...)
}
\arguments{
\item{x}{epanetmsx.rpt object}

\item{elementType}{character indicating whether results for "nodes" or links" should be plotted}

\item{...}{further arguments passed to plotSparklineTable}
}
\description{
Plots a sparkline table of Epanet-msx results
}
\seealso{
plotSparklineTable
}
