% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expandedLinkTable-s3.r
\name{plot.expandedLinkTable}
\alias{plot.expandedLinkTable}
\title{plot an expanded link table}
\usage{
\method{plot}{expandedLinkTable}(x, add = FALSE, label = FALSE,
  linewidths = 3, ...)
}
\arguments{
\item{x}{object of type expandedLinkTable}

\item{add}{logical indicating whether to add this plot to an existing one}

\item{label}{logical indicating if the links should be labeled at the mid points}

\item{linewidths}{passed to lwd argument in segments()}

\item{...}{further arguments passed to plot}
}
\description{
plot an expanded link table
}

