% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sparkline-s3.r
\name{plot.sparkline}
\alias{plot.sparkline}
\title{Plot a sparkline}
\usage{
\method{plot}{sparkline}(x, ...)
}
\arguments{
\item{x}{sparkline object}

\item{...}{further arguments passed to plot.default}
}
\description{
Plot a sparkline
}

