% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparklineTable-s3.r
\name{plot.sparklineTable}
\alias{plot.sparklineTable}
\title{Plot Sparkline Table}
\usage{
\method{plot}{sparklineTable}(x, ...)
}
\arguments{
\item{x}{object of class sparklineTable}

\item{...}{further arguments passed to par}
}
\description{
Plot Sparkline Table
}
