% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.inp-s3.r
\name{plotElementsLegend}
\alias{plotElementsLegend}
\title{Plot Legend of Network Elements}
\usage{
plotElementsLegend(legend.locn)
}
\arguments{
\item{legend.locn}{keyword for location of legend. See details of legend()
function.}
}
\description{
Add legend of network elements to the active plot
}
\details{
Helper function for adding a legend to the active plot.  
Uses plot characters 16, 15, 8 and 25 for Tanks, Reservoirs, Pumps and Valves 
for compatibility with plotInpNodes()
}
\examples{
## make a new blank plot 
plot( c(0,1), c(0,1), type = 'n') 
## add the nodes, including junctions 
plotElementsLegend('topright') 
}
