% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epanet.inp-s3.r
\name{plotInpLegend}
\alias{plotInpLegend}
\title{Plot inp Legend}
\usage{
plotInpLegend(legend.locn)
}
\arguments{
\item{legend.locn}{passed to legend()}
}
\description{
Add a legend to an existing plot
}

