% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epanet.inp-s3.r
\name{plotInpLinks}
\alias{plotInpLinks}
\title{Plot .inp Links}
\usage{
plotInpLinks(x)
}
\arguments{
\item{x}{epanet.inp object}
}
\description{
Add lines for pipes, pumps and valves
 from an epanet.inp object to an existing plot
}

