% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epanet.inp-s3.r
\name{plotInpNodes}
\alias{plotInpNodes}
\title{plot node elements}
\usage{
plotInpNodes(x, plot.junctions)
}
\arguments{
\item{x}{epanet.inp object}

\item{plot.junctions}{logical indicating whether to plot junctions}
}
\description{
Adds node elements from epanet.inp object to an existing plot
}

