% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.inp-s3.r
\name{print.summary.epanet.inp}
\alias{print.summary.epanet.inp}
\title{Print Method for epanet.inp summaries}
\usage{
\method{print}{summary.epanet.inp}(x, ...)
}
\arguments{
\item{x}{object of class summary.epanet.inp}

\item{...}{further arguments}
}
\description{
Prints summary info about the network
}
