% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.rpt-s3.r
\name{print.summary.epanet.rpt}
\alias{print.summary.epanet.rpt}
\title{Print rpt summary}
\usage{
\method{print}{summary.epanet.rpt}(x, ...)
}
\arguments{
\item{x}{a summary.epanet.rpt object}

\item{...}{further arguments passed to print}
}
\description{
The function prints a summary of simulation results
contained in the rpt file.
}
