% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanetmsx.rpt-s3.r
\name{print.summary.epanetmsx.rpt}
\alias{print.summary.epanetmsx.rpt}
\title{Print msx rpt summary}
\usage{
\method{print}{summary.epanetmsx.rpt}(x, ...)
}
\arguments{
\item{x}{a summary.epanetmsx.rpt object}

\item{...}{further arguments passed to print}
}
\description{
The function prints a summary of multi-species simulation results
contained in the report file
}
