% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epanetmsx.rpt-s3.r
\name{read.msxrpt}
\alias{epanetmsx.rpt}
\alias{read.msxrpt}
\title{Read msx results}
\usage{
read.msxrpt(file)
}
\arguments{
\item{file}{the name of the file to read}
}
\value{
Returns an epanet.rpt S3 object with two data.frame elements.

\item{nodeResults}{data.frame}
\item{linkResults}{data.frame}
}
\description{
reads an Epanet-msx .rpt file into R
}
\examples{
# path to Net1.inp example file included with this package
msr <- file.path( find.package("epanetReader"), "extdata","example.rpt")

#read the results into R
x <- read.msxrpt(msr)
names(x)
summary(x)
plot(x)
}

