% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epanet.rpt-s3.r
\name{read.rpt}
\alias{read.rpt}
\title{Read .rpt file}
\usage{
read.rpt(rptFile)
}
\arguments{
\item{rptFile}{the file to read}
}
\value{
a list of data frames
}
\description{
reads an Epanet .rpt file into R
}
\details{
add lines "Page 0", "Links All" and "Nodes All" to the
 [REPORT] section of the .inp file to output info to read in
with this function
}
\references{
Rossman, L. A. (2000). Epanet 2 users manual. US EPA, Cincinnati, Ohio.
}

