% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_file_reader.r
\name{read_lines_wrapper}
\alias{read_lines_wrapper}
\title{Read lines wrapper}
\usage{
read_lines_wrapper(file)
}
\arguments{
\item{file}{the name of the file to read}
}
\value{
character vector where each entry corresponds to 
a line in the file.
}
\description{
Wrapper function for different implementations of
readlines functions
}
\details{
calls Kmisc::readlines if available and base::readLines otherwise
}
