% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparklineTable-s3.r
\name{sparklineTable}
\alias{sparklineTable}
\title{Sparkline Table}
\usage{
sparklineTable(df, row.var, col.vars, xvar = NULL, xrange.labels = NULL)
}
\arguments{
\item{df}{data.frame of values to plot.}

\item{row.var}{variable for rows of the table}

\item{col.vars}{variables for columns of the table}

\item{xvar}{optional name of variable for horizontal axis of sparkline plots}

\item{xrange.labels}{optional vector of length 2 with labels for the first
and last quantities plotted on x-axis, often a date and/or time}
}
\description{
Create S3 object of data for table of sparklines
}
\seealso{
plotSparklineTable
}
