% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.inp-s3.r
\name{summary.epanet.inp}
\alias{summary.epanet.inp}
\title{Summary Method for epanet.inp}
\usage{
\method{summary}{epanet.inp}(object, ...)
}
\arguments{
\item{object}{of class epanet.inp}

\item{...}{futher arguments}
}
\description{
Summarizes the network by printing the Title of the
network and the number of each type of elements.
}
