% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanetmsx.rpt-s3.r
\name{summary.epanetmsx.rpt}
\alias{summary.epanetmsx.rpt}
\title{Summary of Epanet-msx Simulation Results}
\usage{
\method{summary}{epanetmsx.rpt}(object, ...)
}
\arguments{
\item{object}{of epanetmsx.rpt class}

\item{...}{further arguments passed to summary()}
}
\description{
Provides a basic summary of simulation results
}
