% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.inp.r
\name{write.inp}
\alias{write.inp}
\title{Write .inp file}
\usage{
write.inp(x, file)
}
\arguments{
\item{x}{epanet.inp object to write}

\item{file}{the name of the file where object is written}
}
\value{
nothing
}
\description{
Write an epanet.inp object to a file
}
\details{
Writes an epanet.inp object to a file suitable for simulation
with EPANET.
}
\examples{
write.inp(Net1, "Net1-fromR.inp")
n1 <- read.inp("Net1-fromR.inp")
all.equal(Net1, n1)

}
