% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptFuncs.r
\name{binBreaker}
\alias{binBreaker}
\title{Bin Breaker}
\usage{
binBreaker(x, nbin)
}
\arguments{
\item{x}{vector to find cuts for}

\item{nbin}{number of bins}
}
\value{
list with elements Breaks and Labels
}
\description{
Generate break points for use with cut() 
and range labels based on sample max and min
}
\details{
Helpful in making labels use the acutal max and min rather than
the +/- 1% cut() uses by default.
}
