% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanetReader.r
\docType{package}
\name{epanetReader}
\alias{epanetReader}
\alias{epanetReader-package}
\title{Read text files from Epanet into R}
\description{
\pkg{epanetReader} is a package for reading water network information
and simulation results in Epanet's .inp
and .rpt files into R.  See functions \link{read.inp}, 
\link{read.rpt}, and \link{read.msxrpt} for some examples.
}
\author{
Bradley J. Eck
}
