% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandedLinkTable-s3.r
\name{expandedLinkTable}
\alias{expandedLinkTable}
\title{Expanded Link Table}
\usage{
expandedLinkTable(Links, Coordinates)
}
\arguments{
\item{Links}{data frame of Pipes, Pumps or Valves of from epanet.inp}

\item{Coordinates}{table of epanet.inp}
}
\value{
an expandedLinkTable object
}
\description{
Create an expandedLinkTable object by adding node coordinates to a 
data frame of pipes, pumps, or valves.
}
\examples{
x <- expandedLinkTable(Net1$Pipes, Net1$Coordinates) 
print(x)
plot(x) 
}
