% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Net1-data.r
\docType{data}
\name{Net1}
\alias{Net1}
\title{Epanet's Net1 Example}
\format{An object of class \code{epanet.inp} created by \link{read.inp}.}
\source{
http://www.epa.gov/sites/production/files/2014-06/en2setup_0.exe
}
\usage{
Net1
}
\description{
A dataset created by reading the Net1.inp file
distributed with Epanet using this package's
read.inp() function.
}
\examples{
#confirm built-in dataset matches output of read.inp
inp <- file.path( find.package("epanetReader"), "extdata","Net1.inp") 
n1 <- suppressWarnings( read.inp(inp) )
ok <- isTRUE( all.equal(Net1, n1))
if( ok==FALSE) stop("built-in Net1 doesn't match read.inp")
}
