% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Net1rpt-data.r
\docType{data}
\name{Net1rpt}
\alias{Net1rpt}
\title{Epanet's Net1 Example}
\format{An object of class \code{epanet.rpt} created by \link{read.rpt}.}
\usage{
Net1rpt
}
\description{
A dataset created by reading the Net1.rpt file
distributed with Epanet using this package's
read.rpt() function.
}
\examples{
#confirm built-in dataset matches output of read.rpt
rpt <- file.path( find.package("epanetReader"), "extdata","Net1.rpt") 
n1r <- read.rpt(rpt)
ok <- isTRUE( all.equal(Net1rpt, n1r))
if( ok==FALSE) stop("built-in Net1rpt doesn't match read.rpt")
}
