% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.inp-s3.r
\name{plot.epanet.inp}
\alias{plot.epanet.inp}
\title{Plot Method for epanet.inp}
\usage{
\method{plot}{epanet.inp}(x, plot.junctions = TRUE,
  legend.locn = "topright", ...)
}
\arguments{
\item{x}{object of class epanet.inp}

\item{plot.junctions}{logical indicating whether to plot junctions}

\item{legend.locn}{character string passed to legend() specifying
the location of the legend on the plot}

\item{...}{other arguments passed to plot()}
}
\description{
Make a plot of the network using base graphics
}
\details{
Implements the generic plot function for S3 objects of class epanet.inp.
The plot is built from base graphics by creating a blank plot and then calling 
the helper functions plotInpLinks(), plotInpNodes(), plotElementsLegend().
}
\examples{
plot(Net1) 
}

