% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandedLinkTable-s3.r
\name{plot.expandedLinkTable}
\alias{plot.expandedLinkTable}
\title{plot an expanded link table}
\usage{
\method{plot}{expandedLinkTable}(x, add = FALSE, label = FALSE,
  linewidths = 3, color = "black", ...)
}
\arguments{
\item{x}{object of type expandedLinkTable}

\item{add}{logical indicating whether to add to the currently active plot.
add=FALSE creates a new plot.}

\item{label}{logical indicating if the links should be labeled at the mid points}

\item{linewidths}{passed to lwd argument in segments()}

\item{color}{passed to col argument in segments()}

\item{...}{further arguments passed to segments()}
}
\description{
plot an expanded link table
}
\details{
An implementation of the generic plot function for
expandedLinkTable objects. Links are drawn using segments(). 
Useful for building up network plots.
}
