% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline-s3.r
\name{plot.sparkline}
\alias{plot.sparkline}
\title{Plot a sparkline}
\usage{
\method{plot}{sparkline}(x, ...)
}
\arguments{
\item{x}{sparkline object}

\item{...}{further arguments passed to plot.default}
}
\description{
Plot a sparkline
}
\details{
Implementation of the generic plot function for a single sparkline object.
The primarily used to build up plots of a sparklineTable
}
\seealso{
sparkline
}
