% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.inp-s3.r
\name{plotInpLinks}
\alias{plotInpLinks}
\title{Plot .inp Links}
\usage{
plotInpLinks(x, lwd = 3, col = "black")
}
\arguments{
\item{x}{epanet.inp object}

\item{lwd}{width of lines}

\item{col}{color of lines}
}
\description{
Add lines for pipes, pumps and valves
 from an epanet.inp object to an existing plot
}
\details{
Helper function for building up a plot of the network by
adding links to an existing plot.
}
\examples{
## make a new blank plot 
plot( range(Net1$Coordinates$X), range(Net1$Coordinates$Y), type = 'n') 
## add the links
plotInpLinks(Net1) 
}
