% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epanet.inp-s3.r
\name{plotInpNodes}
\alias{plotInpNodes}
\title{Plot Node Elements}
\usage{
plotInpNodes(x, plot.junctions)
}
\arguments{
\item{x}{epanet.inp object}

\item{plot.junctions}{logical indicating whether to plot junctions}
}
\description{
Adds node elements from epanet.inp object to an existing plot
}
\details{
Tanks and Reservoirs are shown using plot characters (pch)
          16 and 15. Junctions, if plotted, appear as pch ="."
}

