% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.inp-s3.r
\name{plotInpNodes}
\alias{plotInpNodes}
\title{Plot Node Elements}
\usage{
plotInpNodes(x, plot.junctions)
}
\arguments{
\item{x}{epanet.inp object}

\item{plot.junctions}{logical indicating whether to plot junctions}
}
\description{
Adds node elements from epanet.inp object to an existing plot
}
\details{
Helper function for building up a network plot. Tanks and
Reservoirs are shown using plot characters (pch) '           16 and 15.
Junctions, if plotted, appear as pch ="."
}
\examples{
## make a new blank plot 
plot( range(Net1$Coordinates$X), range(Net1$Coordinates$Y), type = 'n') 
## add the nodes, including junctions 
plotInpNodes(Net1, TRUE ) 
}
