% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.rpt-s3.r
\name{summary.epanet.rpt}
\alias{summary.epanet.rpt}
\title{Summary of Epanet Simulation Results}
\usage{
\method{summary}{epanet.rpt}(object, ...)
}
\arguments{
\item{object}{of epanet.rpt class}

\item{...}{further arguments passed to summary()}
}
\description{
Provides a basic summary of simulation results
}
\details{
Summary of pipe results shows positive and negative
values of flow but only positive values of velocity
as in the rpt file.
}
